Python transfer, Comment marking disabled spells
